move old specs to opal-spec for now
