+++ opal-spec/callbacks_spec.rb
