+++ opal-spec/component_spec.rb
