+++ opal-spec/element_spec.rb
