+++ opal-spec/react_spec.rb
