+++ opal-spec/reactjs/index.html.erb
@@ -3,7 +3,7 @@
 <head>
 </head>
 <body>
-  <%= javascript_include_tag 'react-with-addons' %>
+  <%= javascript_include_tag 'react' %>
   <%= javascript_include_tag @server.main %>
   <div id="placeholder" style="display: none"></div>
   <div id="render_here"></div>
