+++ opal-spec/tutorial/tutorial_spec.rb
