+++ opal-spec/validator_spec.rb
